+++ core/client/bower.json
@@ -8,9 +8,9 @@
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "0.1.3",
     "ember-data": "1.0.0-beta.18",
-    "ember-mocha": "0.7.0",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-resolver": "0.1.15",
+    "ember-mocha": "0.8.2",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-resolver": "0.1.18",
     "ember-simple-auth": "0.8.0",
     "fastclick": "1.0.6",
     "google-caja": "5669.0.0",
