+++ core/client/tests/unit/controllers/settings-general_test.js
@@ -6,6 +6,9 @@ import {
 
 describeModule(
     'controller:settings/general',
+    {
+        needs: ['service:notifications']
+    },
 
     function () {
         it('isDatedPermalinks should be correct', function () {
